API/requirements.txt
# Core Deep Learning Framework
tensorflow>=2.10.0

# Data Processing and Numerical Computing
numpy>=1.21.0
pandas>=1.3.0

# Image Processing
Pillow>=8.3.0
opencv-python>=4.5.0

# Data Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Jupyter Notebook Environment
jupyter>=1.0.0
notebook>=6.4.0
ipykernel>=6.0.0

# Additional Utilities
scikit-learn>=1.0.0
scipy>=1.7.0

fastapi
uvicorn
python-multipart
tensorflow-serving-api==2.5.0

# Optional: For GPU support (uncomment if using GPU)
# tensorflow-gpu>=2.10.0

# Optional: For advanced image processing
# scikit-image>=0.18.0


